refactor: 스레드, 일정알림 컴포넌트 스레드 리스트 하위로 이동
