+++ frontend/src/components/ThreadList/Thread/Thread.styled.ts
