+++ frontend/src/components/ThreadList/ThreadList.tsx
@@ -2,9 +2,9 @@ import { useFetchThreads } from '~/hooks/queries/useFetchThreads';
 import * as S from './ThreadList.styled';
 import type { ThreadSize } from '~/types/size';
 import { THREAD_TYPE } from '~/constants/feed';
-import Thread from '~/components/Thread/Thread';
+import Thread from '~/components/ThreadList/Thread/Thread';
 import NoticeThread from '~/components/NoticeThread/NoticeThread';
-import Notification from '~/components/Notification/Notification';
+import Notification from '~/components/ThreadList/Notification/Notification';
 import { useRef } from 'react';
 import { useIntersectionObserver } from '~/hooks/useIntersectionObserver';
 import Text from '~/components/common/Text/Text';
